Release candidate for mbed-os-5.4.6

+++ targets/TARGET_Maxim/TARGET_MAX32630/us_ticker.c
@@ -140,7 +140,7 @@ void us_ticker_init(void)
     cfg.compareCount = 0xFFFFFFFF;
     TMR32_Config(US_TIMER, &cfg);
 
-    NVIC_SetVector(US_TIMER_IRQn, tmr_handler);
+    NVIC_SetVector(US_TIMER_IRQn, (uint32_t)tmr_handler);
     NVIC_EnableIRQ(US_TIMER_IRQn);
     TMR32_EnableINT(US_TIMER);
